tools: build `lex` and `yacc` on a non-illumos host
`usr/src/tools/{lex,yacc}` exist to give a cross-build the gate's own
scanner and parser generators, which `Makefile.master` then names as
`$(ONBLD_TOOLS)/bin/$(MACH)/{lex,yacc}`. They do not build on a host
without illumos headers, because both include `<widec.h>`.

`<widec.h>` is `<wchar.h>` plus the EUC codeset machinery and the
pre-standard "wide string" functions. lex and yacc use only five of the
latter -- `wslen`, `wscpy`, `wscmp`, `wsprintf` and `watoi` -- and
otherwise want nothing the host's `<wchar.h>` does not already provide.

Add a `<widec.h>` shim under `cmd/sgs/include`, which both tool Makefiles
already put on the include path as `-I$(SGSDIR)/include`, so no Makefile
changes are needed. It maps those five onto their ISO C equivalents.
`wsprintf` is the only one that is not a rename: it takes a *narrow*
format string and emits wide characters, so `swprintf` is not a drop-in
and the shim formats narrow and widens the result.

Where a real `<widec.h>` exists it is used unchanged, via
`__has_include_next`. That test, rather than `#ifdef __sun`, is what makes
this work: `Makefile.master` passes `-D__sun` to the *native* compiles as
well as the target ones, so `__sun` says nothing about which libc's
headers are in front of the compiler.

+++ usr/src/cmd/sgs/lex/common/ldefs.h
@@ -31,6 +31,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 #include <widec.h>
 #include <wctype.h>
@@ -188,10 +189,10 @@ extern int fptr;
 extern int prev;	/* previous input character */
 extern int pres;	/* present input character */
 extern int peek;	/* next input character */
-extern int *name;
-extern int *left;
-extern int *right;
-extern int *parent;
+extern intptr_t *name;
+extern intptr_t *left;
+extern intptr_t *right;
+extern intptr_t *parent;
 extern Boolean *nullstr;
 extern int tptr;
 extern CHR pushc[TOKENSIZE];
@@ -274,8 +275,8 @@ int cpyact(void);
 int lookup(CHR *s, CHR **t);
 int usescape(int c);
 int alpha(int c);
-int mn2(int a, int d, int c);
-int mn1(int a, int d);
+int mn2(int a, intptr_t d, intptr_t c);
+int mn1(int a, intptr_t d);
 int mn0(int a);
 int dupl(int n);
